+++ tests/RequestTest.php
@@ -89,6 +89,9 @@ class RequestTest extends TestCase
         $this->assertEquals('name=value', $request->getContent());
     }
 
+    /**
+     * @depends testFormatData
+     */
     public function testToString()
     {
         $request = new Request([
@@ -102,7 +105,6 @@ class RequestTest extends TestCase
             'name' => 'value',
         ];
         $request->setData($data);
-        $request->prepare();
 
         $expectedResult = <<<EOL
 POST http://domain.com/test
@@ -111,6 +113,20 @@ Content-Type: application/x-www-form-urlencoded; charset=UTF-8
 name=value
 EOL;
         $this->assertEquals($expectedResult, $request->toString());
+
+        // @see https://github.com/yiisoft/yii2-httpclient/issues/70
+        $request = new Request([
+            'client' => new Client(),
+            'format' => Client::FORMAT_URLENCODED,
+            'method' => 'post',
+            'url' => 'http://domain.com/test',
+        ]);
+        $request->setData($data);
+        $request->addFileContent('some-file', 'some content');
+
+        $result = $request->toString();
+        $this->assertContains('Content-Type: multipart/form-data; boundary=', $result);
+        $this->assertContains('some content', $result);
     }
 
     /**